# specs/gateway.openapi.yaml
openapi: 3.0.0
info:
  title: Secure API Gateway(APIGATE)
  version: 1.0.0
  description: Gateway contract for validated requests.

servers:
  - url: https://localhost:8443
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EchoRequest:
      type: object
      additionalProperties: true
    EchoResponse:
      type: object
      additionalProperties: true
    OrderItem:
      type: object
      required:
        - sku
        - quantity
      properties:
        sku:
          type: string
          minLength: 3
          maxLength: 64
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
    OrderCreate:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: string
          minLength: 3
          maxLength: 64
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        note:
          type: string
          maxLength: 2000
          nullable: true
    OrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
    UserCreate:
      type: object
      required:
        - email
        - display_name
      properties:
        email:
          type: string
          minLength: 5
          maxLength: 255
        display_name:
          type: string
          minLength: 1
          maxLength: 120
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string

paths:
  /v1/echo:
    post:
      summary: Echo payload for testing.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EchoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EchoResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/orders:
    post:
      summary: Create a new order.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '200':
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/orders/{order_id}:
    get:
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users:
    post:
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'